+++ packages/common-ui/styles/common.less
@@ -982,7 +982,7 @@ div.chance-stat-display, div.multiplier-stat-display, div.multiplier-mit-stat-di
 }
 
 
-#new-sheet-form, save-as-modal {
+#new-sheet-form, save-as-modal, change-props-modal {
   fieldset {
     text-align: center;
     display: inline-grid;
