+++ packages/frontend/src/scripts/components/sheet.ts
@@ -50,8 +50,7 @@ import {
     RACE_STATS,
     RaceName,
     STAT_ABBREVIATIONS,
-    SupportedLevel,
-    JOB_DATA
+    SupportedLevel
 } from "@xivgear/xivmath/xivconstants";
 import {getCurrentHash, getCurrentState, processNav} from "../nav_hash";
 import {MateriaTotalsDisplay} from "./materia";
@@ -77,7 +76,7 @@ import {SimCurrentResult, SimResult, SimSettings, SimSpec, Simulation} from "@xi
 import {getRegisteredSimSpecs} from "@xivgear/core/sims/sim_registry";
 import {makeUrl, NavState, ONLY_SET_QUERY_PARAM} from "@xivgear/core/nav/common_nav";
 import {simMaintainersInfoElement} from "./sims";
-import {SaveAsModal, BaseSheetSettingsModal} from "./new_sheet_form";
+import {ChangePropsModal, SaveAsModal} from "./new_sheet_form";
 import {DropdownActionMenu} from "./dropdown_actions_menu";
 import {CustomFoodPopup, CustomItemPopup} from "./custom_item_manager";
 import {confirmDelete} from "@xivgear/common-ui/components/delete_confirm";
@@ -1478,10 +1477,11 @@ export class GearPlanSheetGui extends GearPlanSheet {
 
         const siFmt = formatSyncInfo(this.syncInfo, this.level);
         if (siFmt !== null) {
-            const span = quickElement('span', [], [siFmt]);
-            const ilvlSyncLabel = quickElement('div', ['like-a-button', 'level-sync-info'], [span]);
-            ilvlSyncLabel.title = 'To change the item level sync, click the "Save As" button to create a new sheet with a different level/ilvl.';
-            buttonsArea.appendChild(ilvlSyncLabel);
+            const ilvlSyncBtn = makeActionButton(siFmt, () => {
+                const modal = new ChangePropsModal(this);
+                modal.attachAndShowExclusively();
+            });
+            buttonsArea.appendChild(ilvlSyncBtn);
         }
 
         if (!this.isViewOnly) {
@@ -1533,12 +1533,22 @@ export class GearPlanSheetGui extends GearPlanSheet {
         }
         else {
             sheetOptions.addAction({
-                label: 'Save As',
+                label: 'Save As...',
                 action: () => {
                     const modal = new SaveAsModal(this, newSheet => openSheetByKey(newSheet.saveKey));
                     modal.attachAndShowExclusively();
                 },
             });
+            if (this.saveKey) {
+                sheetOptions.addAction({
+                    label: 'Change Level/Job/Sync',
+                    action: () => {
+                        const modal = new ChangePropsModal(this);
+                        modal.attachAndShowExclusively();
+                    },
+                });
+
+            }
         }
 
         if (!this.isViewOnly) {
@@ -2160,60 +2170,6 @@ export class GearPlanSheetGui extends GearPlanSheet {
         }
         new ChangePropsModal(this).attachAndShowExclusively();
     }
-}
-
-export class ChangePropsModal extends BaseSheetSettingsModal {
-    constructor(private readonly sheet: GearPlanSheetGui) {
-        super({
-            name: sheet.sheetName,
-            job: sheet.classJobName,
-            level: sheet.level,
-            ilvlSyncEnabled: sheet.ilvlSync !== undefined,
-            ilvlSyncLevel: sheet.ilvlSync,
-            allowedRoles: [JOB_DATA[sheet.classJobName].role],
-            multiJob: sheet.isMultiJob,
-        }, 'Apply');
-        this.headerText = 'Change Sheet Properties';
-    }
-
-    protected onSubmit(): void {
-        const desiredJob = this.selectedJob ?? this.sheet.classJobName;
-        const desiredMultiJob = this.multiJob;
-
-        const newName = this.nameValue;
-        const newLevel = this.level;
-        const newIlvl = this.ilvlSyncEnabled ? this.ilvlSync : undefined;
-
-        if (!this.confirmJobMultiChange(this.sheet.classJobName, this.sheet.isMultiJob, desiredJob, desiredMultiJob)) {
-            return;
-        }
-
-        const changed = (this.sheet.sheetName !== newName)
-            || (this.sheet.classJobName !== desiredJob)
-            || (this.sheet.isMultiJob !== desiredMultiJob)
-            || (this.sheet.level !== newLevel)
-            || (this.sheet.ilvlSync !== newIlvl);
-        if (!changed) {
-            this.close();
-            return;
-        }
-
-        // Apply in-place updates for all fields, then save and reload this sheet
-        this.sheet.sheetName = newName;
-        this.sheet.classJobName = desiredJob;
-        this.sheet.isMultiJob = desiredMultiJob;
-        this.sheet.level = newLevel as SupportedLevel;
-        this.sheet.ilvlSync = newIlvl;
-        this.sheet.saveData();
-        if (this.sheet.saveKey) {
-            openSheetByKey(this.sheet.saveKey);
-        }
-        this.close();
-    }
-}
-
-if (!customElements.get('change-props-modal')) {
-    customElements.define('change-props-modal', ChangePropsModal);
 }
 
 export class ImportSetsModal extends BaseModal {
